feat: update DocsLayout
